implemented log function; improved stability; added unit role check
Log function uses images search to record the loot for each battle and save it to a *.csv file. You can choose to enable/disable this function.
More mechanisms to handling errors.
Specify a role for a unit in flow files with DESC entry to enable role check before enter a stage for a first time in the automatic grinding.